Update Ui component
Update Ui component to rename all Person instance to Patient
