+++ src/main/java/seedu/address/ui/PatientCard.java
@@ -12,9 +12,9 @@ import seedu.address.model.patient.Patient;
 /**
  * An UI component that displays information of a {@code Patient}.
  */
-public class PersonCard extends UiPart<Region> {
+public class PatientCard extends UiPart<Region> {
 
-    private static final String FXML = "PersonListCard.fxml";
+    private static final String FXML = "PatientListCard.fxml";
 
     /**
      * Note: Certain keywords such as "location" and "resources" are reserved keywords in JavaFX.
@@ -42,9 +42,9 @@ public class PersonCard extends UiPart<Region> {
     private FlowPane tags;
 
     /**
-     * Creates a {@code PersonCode} with the given {@code Patient} and index to display.
+     * Creates a {@code PatientCode} with the given {@code Patient} and index to display.
      */
-    public PersonCard(Patient patient, int displayedIndex) {
+    public PatientCard(Patient patient, int displayedIndex) {
         super(FXML);
         this.patient = patient;
         id.setText(displayedIndex + ". ");
