+++ src/main/java/seedu/address/ui/PatientListPanel.java
@@ -1,0 +1,49 @@
+package seedu.address.ui;
+
+import java.util.logging.Logger;
+
+import javafx.collections.ObservableList;
+import javafx.fxml.FXML;
+import javafx.scene.control.ListCell;
+import javafx.scene.control.ListView;
+import javafx.scene.layout.Region;
+import seedu.address.commons.core.LogsCenter;
+import seedu.address.model.patient.Patient;
+
+/**
+ * Panel containing the list of patients.
+ */
+public class PatientListPanel extends UiPart<Region> {
+    private static final String FXML = "PatientListPanel.fxml";
+    private final Logger logger = LogsCenter.getLogger(PatientListPanel.class);
+
+    @FXML
+    private ListView<Patient> patientListView;
+
+    /**
+     * Creates a {@code PatientListPanel} with the given {@code ObservableList}.
+     */
+    public PatientListPanel(ObservableList<Patient> patientList) {
+        super(FXML);
+        patientListView.setItems(patientList);
+        patientListView.setCellFactory(listView -> new PatientListViewCell());
+    }
+
+    /**
+     * Custom {@code ListCell} that displays the graphics of a {@code Patient} using a {@code PatientCard}.
+     */
+    class PatientListViewCell extends ListCell<Patient> {
+        @Override
+        protected void updateItem(Patient patient, boolean empty) {
+            super.updateItem(patient, empty);
+
+            if (empty || patient == null) {
+                setGraphic(null);
+                setText(null);
+            } else {
+                setGraphic(new PatientCard(patient, getIndex() + 1).getRoot());
+            }
+        }
+    }
+
+}
