+++ src/main/resources/view/DarkTheme.css
@@ -328,7 +328,7 @@
     -fx-text-fill: white;
 }
 
-#filterField, #personListPanel, #personWebpage {
+#filterField, #patientListPanel, #patientWebpage {
     -fx-effect: innershadow(gaussian, black, 10, 0, 0, 0);
 }
